First Program in Assignmnet-6
Create a simple “shape” hierarchy: a base class called Shape and derived classes called Circle, Square, and Triangle. 
In the base class, make a virtual function called draw( ), and override this in the derived classes.
Make an array of pointers to Shape objects that you create on the heap (and thus perform upcasting of the pointers), and call draw( ) through the base-class pointers, to verify the behaviour of the virtual function.