Second Program in Assignment-7
Write a function template that takes a single type parameter (T) and accepts four function arguments: an array of T, a start index, a Stop index (inclusive), and an optional initial value. The function returns the sum of all array elements in specified range and the initial value. Use the default constructor of T for the default initial value.